MediaPath#toIntMetrics moved to MediaMeter

+++ runtime/servletsrc/com/exedio/cope/pattern/ErrorLog.java
@@ -69,7 +69,7 @@ final class ErrorLog
 		}
 	}
 
-	long get()
+	int get()
 	{
 		return value.get();
 	}
